requirements.txt
Alfred-Workflow==1.24
